Exit when the main process crashes
This patch fixes a rather embarassing mistake I made. The for..in loop that
consumes messages from Kafka and stuffs them into the queue is not protected by
a try / catch because I think it is better to fail fast in such cases and be
restarted by systemd / upstart. Only problem is, we don't actually exit when
the main process crashes, because the child processes are not marked daemonic,
so the interpreter allows the script to continue to run, even though the main
process has died. Marking them as daemons means that when the main process
dies, its children will get killed too, and the script will actually terminate.

Bug: T117994
Change-Id: Ieecadfc95b999132a7a5dc4c227f6eb1cdc080e3