Use ^ and $ while spliting metric value and type
If stats are sent to the enpoint using jsonp then
a jquery callback URL parameter is included.

With the current regex this would make it all the way throguh
and get sent to graphite in many cases as somewhere in the value
there will be a digit followed by a non digit..

Instead the regex should make sure the value is actually a
value exactly as graphite would expect it!

Change-Id: I8f194570fbdb34205b54310e0da73685ea7472c5