Use systemd's process watchdog to trigger restarts
I3ad8bb49ce was supposed to take care of the problem of statsv getting
stuck and not receiving incoming messages, but it didn't. The traceback
in T150359 makes it clear that there is a bug in pykafka that makes it
handle Kafka restarts poorly.

So instead, rely on systemd's process watchdog functionality to make
sure statsv is terminated (and restarted) if no messages are received in
some span of time, determined by the value of 'WatchdogSec=' in the
service unit's file.

See
https://www.freedesktop.org/software/systemd/man/systemd.service.html#WatchdogSec=
and
https://www.freedesktop.org/software/systemd/man/sd_notify.html#
for more info.

Bug: T150359
Change-Id: Ie471fa762b6f52f6114aec36a89094ddc76eb77c